+++ db/migrate/20250627000001_add_notes_to_things.rb
@@ -1,0 +1,5 @@
+class AddNotesToThings < ActiveRecord::Migration[8.1]
+  def change
+    add_column :things, :notes, :text
+  end
+end