+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2025_06_25_000001) do
+ActiveRecord::Schema[8.1].define(version: 2025_06_27_000001) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -85,6 +85,7 @@ ActiveRecord::Schema[8.1].define(version: 2025_06_25_000001) do
     t.text "description"
     t.string "ip_address"
     t.string "name", null: false
+    t.text "notes"
     t.string "owner"
     t.datetime "updated_at", null: false
     t.index ["name"], name: "index_things_on_name"